refactor(server): move the Postgres schema into a shared db module
node_executions is written by the subscriber and read by the API, but the
schema lived in server/api/internal, so the subscriber could only generate
against it by reaching across a module boundary into another module's internal
directory. Nothing else in server/ does that, and Go would reject the
equivalent import outright.

Moves the schema, migrations and Atlas config to server/db as its own module.
Both sqlc configs now point at the same path from the same depth, and neither
reaches into the other module. go:embed cannot traverse .., so the embed and
apply helpers move with the files, which is what makes it a module rather than
a plain directory.

The API depends on it via a replace to ../db, so the Dockerfile copies
server/db before go mod download can resolve it. Its dependency versions are
pinned to the API's; left to tidy they resolved newer and would have bumped the
API's pgx through MVS.

server/db/** is added to both api and subscriber change detection, so a schema
change runs both drift checks rather than only the one that happens to share a
directory with it.

Migrations and atlas.sum move byte-identically; sqlc regenerates both modules
with no diff.
