+++ go.work
@@ -2,5 +2,6 @@ go 1.26.2
 
 use (
 	./server/api
+	./server/db
 	./server/subscriber
 )
