+++ server/api/cmd/dbmigrate/main.go
@@ -27,8 +27,8 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/reearth/reearth-flow/api/internal/infrastructure/mongo/mongodoc"
 	"github.com/reearth/reearth-flow/api/internal/infrastructure/postgres"
-	"github.com/reearth/reearth-flow/api/internal/infrastructure/postgres/db"
 	"github.com/reearth/reearth-flow/api/pkg/id"
+	"github.com/reearth/reearth-flow/db"
 	"github.com/reearth/reearthx/pgxx"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
